[checker][refactor] Move type subst helper into type_system module
